requirements.txt
Django==5.2.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
redis==6.2.0
psycopg==3.2.9
Markdown==3.8.2
django-filter==25.1
django-cors-headers==4.7.0
ruff==0.12.4
ipython==9.4.0
django-environ==0.12.0
simple-ulid==1.0.0
django-phonenumber-field[phonenumbers]==8.1.0
pillow==11.3.0
xlsxwriter==3.2.5
celery[auth,msgpack]==5.5.3
django-celery-beat==2.8.1
sentry-sdk[django]==2.34.1
pytimeparse==1.1.8
duckdb==1.3.2
pytest-django==4.11.1
coverage==7.10.3
pytest-cov==6.2.1
django-prose-editor==0.18.1
django-allauth==65.11.1
mypy==1.17.1
pillow==11.3.0
gunicorn==23.0.0
django-celery-results==2.6.0
django-storages==1.14.6
boto3==1.40.27
go-task-bin==3.44.1